Extract interface from `Messages` class, isolate ES-specific details
This PR is extracting a `MessagesAdapter` interface from the Searches
class, that allows us to separate the Elasticsearch-specific
implementation details from the consumer of that interface. In order to
reuse the mostly implementation-independent `MessagesIT` integration test, it
is made abstract. For every implementation of the `MessagesAdapter`
interface, a class extending the `MessagesIT` class should be
created, handling the `MessagesAdapter`'s implementation-specific
instantiation details.

There are some refactoring possibilities here, namely handling of
success and failures during bulk indexing. There is some domain logic
left inside the helper methodes of the `MessagesAdapterES6`, which
should be moved to the `Messages` class again.

Refs #8155, #8156.